requirements.txt
tensorflow==2.15
matplotlib==3.8
python-dotenv==1.0
labelbox[data]==3.58
boto3==1.34
tqdm==4.66